RP2350: SD-card bootloader fixes
Bootloader was not being included due to too broad linker script wildcards.

Separated ZuluControl platform functions to separate file to make it
easier to exclude from bootloader build.
