+++ utils/run_gdb_picoprobe_rp2350.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 DIR="$( dirname -- "${BASH_SOURCE[0]}"; )";
-ELF="$DIR/../.pio/build/ZuluIDE_RP2350/firmware.elf"
+ELF="$DIR/../.pio/build/ZuluIDE_V2/firmware.elf"
 
 [ -e $ELF ] || pio run
