+++ utils/run_gdb_picoprobe_rp2350_debug.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 DIR="$( dirname -- "${BASH_SOURCE[0]}"; )";
-ELF="$DIR/../.pio/build/ZuluIDE_RP2350_core1_debug/firmware.elf"
+ELF="$DIR/../.pio/build/ZuluIDE_V2_core1_debug/firmware.elf"
 
 [ -e $ELF ] || pio run
